--- Client.py
@@ -1,155 +1,0 @@
-# import pygame
-# from pygame.locals import *
-# from Network import Network
-
-# width = 500
-# height = 500
-# win = pygame.display.set_mode((width, height))
-# pygame.display.set_caption("Client")
-
-# clientNumber = 0
-
-# class Player():
-#     def __init__(self, x, y, width, height, color):
-#         self.x = x
-#         self.y = y
-#         self.width = width
-#         self.height = height
-#         self.color = color
-#         self.rect = (x,y,width,height)
-#         self.vel = 3
-
-#     def draw(self, win):
-#         pygame.draw.rect(win, self.color, self.rect)
-
-#     def move(self):
-#         keys = pygame.key.get_pressed()
-
-#         if keys[pygame.K_LEFT]:
-#             self.x -= self.vel
-
-#         if keys[pygame.K_RIGHT]:
-#             self.x += self.vel
-
-#         if keys[pygame.K_UP]:
-#             self.y -= self.vel
-
-#         if keys[pygame.K_DOWN]:
-#             self.y += self.vel
-
-#         self.update()
-
-#     def update(self):
-#         self.rect = (self.x, self.y, self.width, self.height)
-
-
-# def read_pos(string): # Reads the position sent as a string, splits it into two and changes them into a tuple.
-#     string = string.split(",")
-#     return int(str[0]), int(str[1])
-
-# def make_pos(tup):
-#     return str(tup[0]) + "," + str(tup[1])
-
-# def redrawWindow(win,player, player2):
-#     win.fill((255,255,255))
-#     player.draw(win)
-#     player2.draw(win)
-#     pygame.display.update()
-
-
-# def main():
-#     run = True
-#     n = Network()
-#     startPos = read_pos(n.getPos())
-#     p = Player(startPos[0],startPos[1],100,100,(0,255,0))
-#     p2 = Player(0, 0 ,100,100,(0,255,0))
-#     clock = pygame.time.Clock()
-
-#     while run:
-#         clock.tick(60)
-
-#         p2Pos = read_pos(n.send(make_pos((p.x, p.y)))) # Handles all of Player 2, gets the postion and sends it to server.
-#         p2.x = p2Pos[0]
-#         p2.y = p2Pos[1]
-#         p2.update()
-
-#         for event in pygame.event.get():
-#             if event.type == pygame.QUIT:
-#                 run = False
-#                 pygame.quit()
-
-#         p.move()
-#         redrawWindow(win, p, p2)
-
-# main()
-
-
-# in a server-client network, the client is the thing that processes and sends requests to other clients. This is basically like a "user" or "person"
-# on the network.
-
-# this includes the basic code that is the game.
-
-import pygame
-from Network import Network
-
-width = 500
-height = 500
-win = pygame.display.set_mode((width, height))
-pygame.display.set_caption("Client")
-
-clientNumber = 0
-
-
-class Player():
-    def __init__(self, x, y, width, height, color):
-        self.x = x
-        self.y = y
-        self.width = width
-        self.height = height
-        self.color = color
-        self.rect = (x,y,width,height)
-        self.vel = 3
-
-    def draw(self, win):
-        pygame.draw.rect(win, self.color, self.rect)
-
-    def move(self):
-        keys = pygame.key.get_pressed()
-
-        if keys[pygame.K_LEFT]:
-            self.x -= self.vel
-
-        if keys[pygame.K_RIGHT]:
-            self.x += self.vel
-
-        if keys[pygame.K_UP]:
-            self.y -= self.vel
-
-        if keys[pygame.K_DOWN]:
-            self.y += self.vel
-
-        self.rect = (self.x, self.y, self.width, self.height)
-
-
-def redrawWindow(win,player):
-    win.fill((255,255,255))
-    player.draw(win)
-    pygame.display.update()
-
-
-def main():
-    run = True
-    p = Player(50,50,100,100,(0,255,0))
-    clock = pygame.time.Clock()
-
-    while run:
-        clock.tick(60)
-        for event in pygame.event.get():
-            if event.type == pygame.QUIT:
-                run = False
-                pygame.quit()
-
-        p.move()
-        redrawWindow(win, p)
-
-main()
